로비 채팅창 아이디 추가 + 서버연동 완료
로비 채팅창 아이디 추가 + 서버연동 완료
방만들기를 다시 구현해야함.

+++ Lunatic_Project/Lunatic Project/LunaticProject/HeroManager.h
@@ -35,7 +35,7 @@ public:
 	~CHeroManager();
 
 	int GetHp() { return hp; }
-	float GetAttack() { return attack; }
+	int GetAttack() { return attack; }
 	float GetSpeed() { return speed; }
 	float GetNormalSpeed() { return normalspeed; }
 	float GetRootSpeed() { return rootspeed; }
